Change a yarn integration test to use concurrently instead of npm-run-all to address a Travis build issue.
The issue is
https://travis-ci.org/srs/gradle-node-plugin/builds/177526145. This issue
occurred because yarn deleted the npm locally installed by npmSetup during
the yarn task and so npm-run-all couldn't find npm executable. It could be
fixed by adding devDependency to npm, but instead I replaced npm-run-all
with concurrently because npm should not be used when yarn is being used.