Don't track yarn working directory as an input.
There's actually no need to track the yarn working directory as an input.
The only thing this affects is how we calculate the output directory. Since
we track the output directory explicitly this is unneeded. Removing this
also removes an absolute path from the task inputs, allowing the Gradle
build cache to work across different machines.